modals moved to modals folder & add styling
