+++ src/Timer.js
@@ -1,14 +1,3 @@
-// import React from 'react';
-//
-// const Timer = props => (
-//   <div className={props.seconds > 90 ? 'timer hidden' : 'timer visible'}>
-//     <span>{props.seconds}</span>
-//   </div>
-// );
-//
-// export default Timer;
-
-
 import React, { Component } from 'react';
 
 export default class Timer extends Component {
